+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 - Update the DNS client to follow configured timeouts in a more predictable manner.  Also fix a corner case in its
   behavior that could cause it to resolve incorrectly during transient network and DNS server failures.
   [#11386](https://github.com/Kong/kong/pull/11386)
+- Fix cache warmup mechanism not working in `acls` plugin groups config entity scenario.
+  [#11414](https://github.com/Kong/kong/pull/11414)
 
 #### Plugins